+++ rstar/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 - Add optional support for the [mint](https://docs.rs/mint/0.5.9/mint/index.html) crate
+- Added cached envelope bulk load benchmark
 
 ## Changed
 - Fixed a stack overflow error in `DrainIterator::next`